requirements.txt
blinker==1.9.0  
click==8.2.1
colorama==0.4.6
Flask==3.1.2
Flask-Login==0.6.3
Flask-WTF==1.2.2
hashids==1.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.0.29
pip==25.2
protobuf==6.32.0
python-dotenv==1.1.1
Werkzeug==3.1.3
WTForms==3.2.1
